Add initial zabbix support
This is a very simple change (and one of the first times I've ever touched golang) that outputs numeric (float)s to add to Zabbix to check if a grade is high enough for a certain host. No configuration, just have to change the case depending on what you consider as high enough or not. Add triggers based on values (1=ok,0=bad,-1=untrusted,-1337=error).